backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
chromadb==0.5.15
openai==1.51.0
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.12
pydantic==2.9.2
httpx==0.27.2
